% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStock.R
\name{getStock}
\alias{getStock}
\title{Get Stock}
\usage{
getStock(stock = NULL, year = NULL)
}
\arguments{
\item{stock}{the stock code, e.g. \code{"cod-347d"}.}

\item{year}{the active year of the stock list, e.g. \code{2016}.}
}
\value{
A data frame containing the stock list data.
}
\description{
Get stock list data for one or more stocks, in one or more years.
}
\details{
Pass \code{stock = NULL} to get all stocks, or \code{year = NULL} to get all
years.
}
\note{
The stock list data specify the expert group, advice group, assessment type,
etc.
}
\examples{
getStock("sai-icel")
getStock("cod-347d", 2016)

}
\seealso{
\code{\link{getSD}} shows stock list data for all stocks in all years.

\code{\link{showStock}} shows stock list data for one stock, in a readable
format.

\code{\link{icesSD-package}} gives an overview of the package.
}
