% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icesSD-package.R
\docType{package}
\name{icesSD-package}
\alias{icesSD-package}
\alias{icesSD}
\title{Stock Database Web Services}
\description{
R interface to access the web services of the ICES Stock Database.
}
\details{
\emph{Get dataset:}
\tabular{ll}{
  \code{\link{getSD}}    \tab full stock list\cr
}
}
\references{
ICES Stock Database: \url{http://sd.ices.dk}.

ICES Stock Database web services: \url{http://sd.ices.dk/services/}.
}
\author{
Colin Millar, Scott Large, and Arni Magnusson.
}
