% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icesSLD-package.R
\docType{package}
\name{icesSLD-package}
\alias{icesSLD-package}
\alias{icesSLD}
\title{Stock Database Web Services}
\description{
R interface to access the web services of the ICES Stock Database.
}
\details{
\emph{Get dataset:}
\tabular{ll}{
  \code{\link{getSLD}}    \tab full stock list\cr
  \code{\link{showStock}} \tab one stock
}
}
\references{
ICES Stock Database: \url{http://sd.ices.dk}.

ICES Stock Database web services: \url{http://sd.ices.dk/services/}.
}
\author{
Colin Millar, Scott Large, and Arni Magnusson.
}
