% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_api.R
\name{sd_api}
\alias{sd_api}
\title{Build a Stock Database (SD) web service url}
\usage{
sd_api(service, ...)
}
\arguments{
\item{service}{the name of the service}

\item{...}{name arguments will be added as queries}
}
\value{
a complete url as a character string
}
\description{
utility to build a url with optional query arguments
}
\examples{

sd_api("hi", bye = 21)
sd_api("odata4/StockListDWs4")

}
