% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSD.R
\name{getSD}
\alias{getSD}
\title{Get Stock List Data}
\usage{
getSD()
}
\value{
A data frame.

\code{\link{showStock}} shows stock list data for one stock.

\code{\link{icesSD-package}} gives an overview of the package.
}
\description{
Get stock list data such as stock code, expert group, assessment type, etc.
}
\examples{
sldata <- getSLD()

}
