% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSLD.R
\name{getSLD}
\alias{getSLD}
\title{Get Stock List Data}
\usage{
getSLD()
}
\value{
A data frame.

\code{\link{icesSLD-package}} gives an overview of the package.
}
\description{
Get haul data such as position, depth, sampling method, etc.
}
\examples{
sldata <- getSLD()
str(sldata)

}

