% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSLD.R
\name{getSLD}
\alias{getSLD}
\title{Get Stock List Data}
\usage{
getSLD()
}
\value{
A data frame.

\code{\link{showStock}} shows stock list data for one stock.

\code{\link{icesSLD-package}} gives an overview of the package.
}
\description{
Get stock list data such as stock code, expert group, assessment type, etc.
}
\examples{
sldata <- getSLD()

}
