% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icesSD-package.R
\docType{package}
\name{icesSD-package}
\alias{icesSD-package}
\alias{icesSD}
\title{Stock Database Web Services}
\description{
R interface to access the web services of the ICES Stock Database.
}
\details{
\emph{Get dataset:}
\tabular{ll}{
  \code{\link{getSD}}    \tab full stock list\cr
  \code{\link{getStock}} \tab selected stock(s) and year(s)
}
\emph{Show on screen:}
\tabular{ll}{
  \code{\link{showStock}} \tab one stock, in a readable format
}
}
\references{
ICES Stock Database: \url{https://sd.ices.dk}.

ICES Stock Database web services: \url{https://sd.ices.dk/services/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sd.ices.dk}
  \item \url{https://github.com/ices-tools-prod/icesSD}
  \item Report bugs at \url{https://github.com/ices-tools-prod/icesSD/issues}
}

}
\author{
Colin Millar, Scott Large, and Arni Magnusson.
}
