% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icesSLD-package.R
\docType{package}
\name{icesSLD-package}
\alias{icesSLD}
\alias{icesSLD-package}
\title{Stock List Database Web Services}
\description{
R interface to access the web services of the ICES (International
Council for the Exploration of the Sea) Stock List database.
}
\author{
Colin Millar, Scott Large, and Arni Magnusson.
}
\references{
ICES SLD web services:
\url{http://ices.dk/marine-data/tools/Pages/WebServices.aspx}.

ICES software applications:
\url{http://ices.dk/marine-data/tools/Pages/Software.aspx}.
}

