% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showStock.R
\name{showStock}
\alias{showStock}
\title{Show Stock}
\usage{
showStock(stock = NULL, year = NULL)
}
\arguments{
\item{stock}{the stock code, e.g. cod-347d.}

\item{year}{the active year of the stock list, e.g. 2016, or NULL to get the
most recent year available.}
}
\value{
A single-row data frame, printed to screen with
\code{print.simple.list} unless quiet is TRUE.

\code{\link{getSD}} gets all stock list data.

\code{\link{icesSD-package}} gives an overview of the package.
}
\description{
Show stock list data for a given stock, such as expert group, advice group,
assessment type, etc.
}
\examples{
showStock("sai-icel")
cod.347d <- showStock("cod-347d")

}
